split rtm controller for swiftlint
